feat(algo): fix lint, move invalid const to coin model
