+++ packages/blockchain-wallet-v4-frontend/src/data/components/borrow/selectors.ts
@@ -1,9 +1,7 @@
-import { INVALID_COIN_TYPE } from './model'
+import { getRatesSelector } from 'core/redux/data/misc/selectors'
 import { RatesType } from './types'
 import { RemoteDataType } from 'core/types'
 import { RootState } from '../../rootReducer'
-import { selectors } from 'data'
-import Remote from 'blockchain-wallet-v4/src/remote/remote'
 
 export const getCoinType = (state: RootState) => state.components.borrow.coin
 
@@ -28,14 +26,7 @@ export const getRates = (
 ): RemoteDataType<string | Error, RatesType> => {
   const coinType = getCoinType(state)
 
-  switch (coinType) {
-    case 'BTC':
-      return selectors.core.data.btc.getRates(state)
-    case 'PAX':
-      return selectors.core.data.eth.getErc20Rates(state, 'pax')
-    default:
-      throw Remote.Failure(INVALID_COIN_TYPE)
-  }
+  return getRatesSelector(coinType, state)
 }
 
 export const getStep = (state: RootState) => state.components.borrow.step
