+++ packages/blockchain-wallet-v4-frontend/src/data/components/interest/sagas.utils.ts
@@ -11,7 +11,7 @@ import {
 } from 'core/types'
 import { convertBaseToStandard } from '../exchange/services'
 import { Exchange } from 'blockchain-wallet-v4/src'
-import { INVALID_COIN_TYPE } from './model'
+import { INVALID_COIN_TYPE } from 'core/model'
 import { promptForSecondPassword } from 'services/SagaService'
 import { RatesType } from '../borrow/types'
 import { selectors } from 'data'
