+++ packages/blockchain-wallet-v4/src/model.ts
@@ -1,0 +1,1 @@
+export * from './network/api/coin/model'
