+++ packages/blockchain-wallet-v4/src/network/api/coin/model.ts
@@ -1,0 +1,1 @@
+export const INVALID_COIN_TYPE = 'Invalid coin type'
