--- packages/blockchain-wallet-v4/src/redux/data/misc/selectors.js
@@ -1,20 +1,0 @@
-import { dataPath } from '../../paths'
-import { path } from 'ramda'
-
-export const getCaptcha = path([dataPath, 'misc', 'captcha'])
-
-export const getPriceIndexSeries = path([
-  dataPath,
-  'misc',
-  'price_index_series'
-])
-
-export const getLogs = path([dataPath, 'misc', 'logs'])
-
-export const getPairingCode = path([dataPath, 'misc', 'pairing_code'])
-
-export const authorizeLogin = path([dataPath, 'misc', 'authorize_login'])
-
-export const handle2FAReset = path([dataPath, 'misc', 'handle_2fa_reset'])
-
-export const verifyEmailToken = path([dataPath, 'misc', 'verify_email_token'])
